Changes header and creation Card-Component
