+++ beautifyberlin/src/App.css
@@ -1,62 +1,6 @@
-
-
-.top {
-  width: 100%;
-  height: 80px;
-  background-color: #8242c2;
-  position: sticky;
-  top:0;
-  display: flex;
-  align-items: center;
-}
-
-.topLeft, .topRight, .topCenter {
-  text-align: center;
-  display: flex;
-  justify-content: center;
-}
-
-.topLeft {
-  flex: 4;
-  font-family: 'Pacifico', cursive;
-  color: white;
-  font-size: 40px;
-  cursor: pointer;
-}
-
-.topCenter {
-  flex: 6;
-  font-family: 'Lora', serif;
-}
-
-.topRight {
-  flex: 2;
-  justify-content: right
-}
-
-.topLeft, .topRight {
-  background-color: #8242c2;
-}
-
-.topIcon {
-  font-size: 20px;
-  margin-right: 10px;
-  cursor: pointer;
-  color: white
-}
-
-.topList {
-  display: flex;
-  justify-content: center;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.topListItem {
-  margin-right: 25px;
-  font-size: 18px;
-  font-weight: 300;
-  cursor: pointer;
-  color: white
+.Post{
+   display: flex;
+   padding: 20px;
+   margin: 10px;
+   display: inline-block;
 }
