+++ beautifyberlin/src/App.js
@@ -1,5 +1,6 @@
-import './App.css';
-import TopBar from './TopBar.js';
+import './App.css'
+import TopBar from './components/topbar/TopBar';
+import HOFCard from './components/HOFCard/HOFCard';
 
 function App() {
   return (
@@ -7,6 +8,62 @@ function App() {
       <>
       <TopBar />
       <p>Text</p>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
+      <div className="Post">
+      <HOFCard 
+      title="Seumestraße, Berlin"
+      imageUrl="https://i.pinimg.com/originals/f7/1d/68/f71d68249b385b0554a52e7fbfec13d9.jpg"
+      body="This work of art was made by Ms. Mila Mila. She's one of the most famous artist's in the world."
+      />
+      </div>
       </>
     </div>
   );
